requirements.txt
ozpy
python-ldap
